aula 03 - Interface Globa, Componentes Time e Colaborador
Criação de uma interface para ser utilizada globalmente e refatoração dos componentes Time e Colaborador
