--- Alura/React JS/7 - React - Migrando para Typescript/organo/src/componentes/Colaborador/index.js
@@ -1,17 +1,0 @@
-import "./Colaborador.css"
-
-const Colaborador = ({nome, cargo, imagem, corDeFundo}) => {
-  return (
-    <div className="colaborador">
-      <div className="cabecalho" style={{backgroundColor: corDeFundo}}>
-        <img src={imagem} alt={nome} />
-      </div>
-      <div className="rodape">
-        <h4>{nome}</h4>
-        <h5>{cargo}</h5>
-      </div>
-    </div>
-  )
-}
-
-export default Colaborador
