+++ Alura/React JS/7 - React - Migrando para Typescript/organo/src/componentes/Colaborador/index.tsx
@@ -1,0 +1,34 @@
+import { IColaborador } from "../../shared/interfaces/IColaborador"
+import "./Colaborador.css"
+
+interface ColaboradorProps extends IColaborador{
+  corDeFundo: string;
+}
+
+// interface ColaboradorProps {
+//   colaborador: IColaborador;
+//   corPrimaria: string;
+// }
+
+// interface ColaboradorProps {
+//   nome: string;
+//   imagem: string;
+//   cargo: string;
+//   corDeFundo:string;
+// }
+
+const Colaborador = ({nome, cargo, imagem, corDeFundo}: ColaboradorProps) => {
+  return (
+    <div className="colaborador">
+      <div className="cabecalho" style={{backgroundColor: corDeFundo}}>
+        <img src={imagem} alt={nome} />
+      </div>
+      <div className="rodape">
+        <h4>{nome}</h4>
+        <h5>{cargo}</h5>
+      </div>
+    </div>
+  )
+}
+
+export default Colaborador
