+++ Alura/React JS/7 - React - Migrando para Typescript/organo/src/shared/interfaces/IColaborador.ts
@@ -1,0 +1,5 @@
+export interface IColaborador {
+  nome: string;
+  cargo: string;
+  imagem: string;
+}
